+++ src/main/java/me/afk/librarianRoller/RollerPhaseBreak.java
@@ -1,7 +1,7 @@
 package me.afk.librarianRoller;
 
 import me.afk.librarianRoller.config.ModConfig;
-import me.afk.librarianRoller.utils.MessageUtils;
+import me.afk.librarianRoller.dataModel.VillagerAndLectern;
 import me.afk.librarianRoller.utils.PlayerInventoryUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
@@ -9,8 +9,6 @@ import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.InteractionHand;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.item.AxeItem;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.LecternBlock;
 
@@ -25,7 +23,6 @@ public class RollerPhaseBreak implements IRollerPhase{
     @Override
     public void doAction(RollerContext ctx) {
         if (!ctx.getEnabled()) return;
-        //fixme Even when call stop method, it couldn't stop immediately, especially for an axe with an efficiency of 7.it still break lectern
 
         Minecraft instance = ctx.getMinecraft();
         LocalPlayer player = instance.player;