+++ src/main/java/me/afk/librarianRoller/RollerPhaseInteract.java
@@ -1,7 +1,6 @@
 package me.afk.librarianRoller;
 
-import me.afk.librarianRoller.config.ModConfig;
-import me.afk.librarianRoller.utils.MessageUtils;
+import me.afk.librarianRoller.dataModel.VillagerAndLectern;
 import me.afk.librarianRoller.utils.PlayerInventoryUtils;
 import me.afk.librarianRoller.utils.VillagerUtils;
 import net.minecraft.client.Minecraft;
@@ -10,9 +9,12 @@ import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.npc.villager.Villager;
-import net.minecraft.world.item.AxeItem;
+//? if > 1.21.1 {
+/*import net.minecraft.world.entity.npc.villager.Villager;
+*///?} else {
+
+import net.minecraft.world.entity.npc.Villager;
+        //?}
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
@@ -31,35 +33,44 @@ public class RollerPhaseInteract implements IRollerPhase{
 
 //        Minecraft instance = Minecraft.getInstance();
         Minecraft instance = ctx.getMinecraft();
+        if (instance == null) return;
         LocalPlayer player = instance.player;
-        MultiPlayerGameMode interactionManager = instance.gameMode;
+        MultiPlayerGameMode gameMode = instance.gameMode;
         Level level = instance.level;
-        int pairIndex = ctx.getPairIndex();
+        if (player == null || gameMode == null || level == null) return;
         List<VillagerAndLectern> list = ctx.getList();
-//        ModConfig modConfig = ctx.getModConfig();
+        if (list.isEmpty()) return;
+        int pairIndex = ctx.getPairIndex();
+        if (pairIndex >= list.size()) {
+            ctx.setPairIndex(0);
+            return;
+        }
+        //        ModConfig modConfig = ctx.getModConfig();
 
-        if (player == null || level == null || list.isEmpty() || interactionManager == null) return;
 
         // two case if result is not success
         // 1. the villager is unemploy but lectern exit
         // 2. the lectern is gone
-        BlockPos blockPos = list.get(pairIndex).lecternPos();
+        VillagerAndLectern villagerAndLectern = list.get(pairIndex);
+        if (villagerAndLectern == null) return;
+        BlockPos blockPos = villagerAndLectern.lecternPos();
+        if (blockPos == null )return;
         Block block = level.getBlockState(blockPos).getBlock();
         if (block == Blocks.AIR) {
             ctx.setRollerPhase(RollerPhasePlace.INSTANCE);
+            return;
         }
 
-        Villager villager = list.get(pairIndex).villager();
-        if (!VillagerUtils.isLibrarian(villager)) return;
+        Villager villager = villagerAndLectern.villager();
+        if (villager == null || !VillagerUtils.isLibrarian(villager)) return;
 
         if (PlayerInventoryUtils.swapAxe(ctx, player)) return;
 
-        InteractionResult result = interactionManager.interact(player, villager, InteractionHand.MAIN_HAND);
+        InteractionResult result = gameMode.interact(player, villager, InteractionHand.MAIN_HAND);
         player.swing(InteractionHand.MAIN_HAND);
 
-        /*if (result == InteractionResult.SUCCESS) {
-            ctx.setRollerPhase(RollerPhaseBreak.INSTANCE);
-            return;
-        }*/
+        if (result == InteractionResult.SUCCESS) {
+            ctx.setRollerPhase(RollerPhaseParse.INSTANCE);
+        }
     }
 }