+++ src/main/java/me/afk/librarianRoller/RollerPhaseParse.java
@@ -1,15 +1,19 @@
 package me.afk.librarianRoller;
 
-import me.afk.librarianRoller.config.ModConfigManager;
-import me.afk.librarianRoller.dataModel.EnchantBook;
+import me.afk.librarianRoller.dataModel.Enchantment;
+import me.afk.librarianRoller.dataModel.OfferData;
 import me.afk.librarianRoller.dataModel.MerchantTradeData;
+import me.afk.librarianRoller.utils.EnchantedBookUtils;
 import me.afk.librarianRoller.utils.MessageUtils;
-import net.minecraft.core.Holder;
-import net.minecraft.core.component.DataComponents;
+        //? if >= 1.21.1 {
+        //?} else >= 1.20.1 {
+
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.ListTag;
+import net.minecraft.nbt.Tag;
+        //?}
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.item.enchantment.Enchantment;
-import net.minecraft.world.item.enchantment.ItemEnchantments;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,7 +41,7 @@ public class RollerPhaseParse implements IRollerPhase{
         if (match != null) {
             ctx.setEnchantBook(match);
             MessageUtils.printReward(match.name(), match.level(), match.cost());
-            if (ModConfigManager.INSTANCE.getConfig().autoBuy) {
+            if (ctx.getModConfigManager().getConfig().autoBuy) {
                 ctx.setRollerPhase(RollerPhaseBuy.INSTANCE);
             } else {
                 ctx.stop();
@@ -47,29 +51,51 @@ public class RollerPhaseParse implements IRollerPhase{
         }
     }
 
-    private EnchantBook findMatch(List<MerchantTradeData.SingleTradeEntry> offers, Map<String, Integer> tradEntry) {
+    //? if >= 1.21.1 {
+    /*private OfferData findMatch(List<MerchantTradeData.SingleTradeEntry> offers, Map<String, Integer> tradEntry) {
         //fixme some version like 1.20.1 use another way but datacomponents
 
         for (int i = 0; i < offers.size(); ++i) {
             MerchantTradeData.SingleTradeEntry singleTradeEntry = offers.get(i);
             ItemStack result = singleTradeEntry.result();
+            List<Enchantment> enchantments = EnchantedBookUtils.readStoredEnchantments(result);
 
-            if (!result.is(Items.ENCHANTED_BOOK)) continue;
-            ItemEnchantments enchantments = result.get(DataComponents.STORED_ENCHANTMENTS);
-            if (enchantments == null || enchantments.isEmpty()) continue;
-
-            for (Map.Entry<Holder<Enchantment>, Integer> entry : enchantments.entrySet()) {
-                String name = entry.getKey().value().description().getString().trim().toLowerCase();
-                int lvl = entry.getValue();
-                LOGGER.info("THERE ARE ENCHENMENTS: {}{}", name, lvl);
+            for (var book : enchantments) {
+                int lvl = book.level();
+                String name = book.name();
+                LOGGER.info("THERE ARE ENCHENMENTS: {} {}", name, lvl);
 
                 Integer requiredLevel = tradEntry.get(name);
                 if (requiredLevel != null && lvl >= requiredLevel) {
-                    return new EnchantBook(name, lvl, singleTradeEntry.costA().getCount(), i);
+                    return new OfferData(name, lvl, singleTradeEntry.costA().getCount(), i);
                 }
             }
         }
 
         return null;
     }
+*///?} else {
+private OfferData findMatch(List<MerchantTradeData.SingleTradeEntry> offers, Map<String, Integer> tradEntry) {
+    for (int i = 0; i < offers.size(); ++i) {
+        MerchantTradeData.SingleTradeEntry singleTradeEntry = offers.get(i);
+        ItemStack result = singleTradeEntry.result();
+
+
+        List<Enchantment> enchantedBooks = EnchantedBookUtils.readStoredEnchantments(result);
+
+        for (var book : enchantedBooks) {
+            int lvl = book.level();
+            String enchId = book.name();
+            LOGGER.info("THERE ARE ENCHENMENTS: {} {}", enchId, lvl);
+
+            Integer requiredLevel = tradEntry.get(enchId);
+            if (requiredLevel != null && lvl >= requiredLevel) {
+                // costA count, trade index i
+                return new OfferData(enchId, lvl, singleTradeEntry.costA().getCount(), i);
+            }
+        }
+    }
+    return null;
+}
+    //?}
 }