+++ src/main/java/me/afk/librarianRoller/config/ModeVType.java
@@ -1,0 +1,6 @@
+package me.afk.librarianRoller.config;
+
+public enum ModeVType {
+    CONTINUE,
+    SINGLE
+}