+++ src/main/java/me/afk/librarianRoller/dataModel/EnchantBook.java
@@ -1,4 +1,0 @@
-package me.afk.librarianRoller.dataModel;
-
-public record EnchantBook(String name, int level, int cost, int index) {
-}