+++ src/main/java/me/afk/librarianRoller/dataModel/OfferData.java
@@ -1,0 +1,4 @@
+package me.afk.librarianRoller.dataModel;
+
+public record OfferData(String name, int level, int cost, int index) {
+}