+++ src/main/java/me/afk/librarianRoller/dataModel/VillagerAndLectern.java
@@ -1,11 +1,11 @@
-package me.afk.librarianRoller;
+package me.afk.librarianRoller.dataModel;
 
 import net.minecraft.core.BlockPos;
         //? if >=1.21.11 {
-import net.minecraft.world.entity.npc.villager.Villager;
-        //?} else {
-/*import net.minecraft.world.entity.npc.Villager;
-*///?}
+/*import net.minecraft.world.entity.npc.villager.Villager;
+        *///?} else {
+import net.minecraft.world.entity.npc.Villager;
+//?}
 import net.minecraft.world.phys.BlockHitResult;
 
 public record VillagerAndLectern(Villager villager, BlockPos lecternPos, BlockHitResult lecternBelowHitResult) {