+++ src/main/java/me/afk/librarianRoller/utils/PlayerInventoryUtils.java
@@ -1,5 +1,8 @@
 package me.afk.librarianRoller.utils;
 
+import me.afk.librarianRoller.LibrarianRoller;
+import me.afk.librarianRoller.RollerContext;
+import me.afk.librarianRoller.config.ModConfig;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import net.minecraft.client.player.LocalPlayer;
@@ -8,41 +11,129 @@ import net.minecraft.core.Direction;
 import net.minecraft.network.protocol.game.ServerboundPlayerActionPacket;
 import net.minecraft.network.protocol.game.ServerboundSetCarriedItemPacket;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.ClickType;
+import net.minecraft.world.item.AxeItem;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 
 import java.util.function.Predicate;
 
 public class PlayerInventoryUtils {
+    private static final Logger LOGGER = LoggerFactory.getLogger("PlayerInventoryUtils-" + LibrarianRoller.MOD_ID);
 
-    public static boolean swapItem(LocalPlayer player, EquipmentSlot handSlot, Predicate<Item> predicate) {
-        boolean state = false;
-        if (!predicate.test(player.getItemBySlot(handSlot).getItem())) {
-            MultiPlayerGameMode interactionManager = Minecraft.getInstance().gameMode;
-            if (interactionManager == null) return state;
+    private static final int PLAYER_INV_STORAGE_START_SLOT = 9;
+    private static final int PLAYER_INV_STORAGE_END_SLOT = 45;
+    private static final int PLAYER_INV_HOTBAR_CONTAINER_START = 36;
+    private static final int PLAYER_INV_OFFHAND_CONTAINER_SLOT = 45;
+    private static final int HOTBAR_SIZE = 9;
 
-            if (player.inventoryMenu == player.containerMenu) {
-                state = swapFromInventory(player, interactionManager, predicate, handSlot);
-            } else {
-                state = swapFromHotbar(player, handSlot, predicate);
+    public static boolean swapAxe(RollerContext ctx, LocalPlayer player) {
+        if (!(PlayerInventoryUtils.swapItem(player, EquipmentSlot.MAINHAND, itemHere -> itemHere instanceof AxeItem))){
+            MessageUtils.throwError("afk.enchant_roller.error.not_found_axe");
+            ctx.stop();
+
+            return true;
+        }
+        return false;
+    }
+
+    public static boolean preventAxeBreaking(RollerContext ctx, LocalPlayer player, ModConfig modConfig) {
+        boolean bl = false;
+
+        var stack = player.getMainHandItem();
+
+        if (!(stack.getItem() instanceof AxeItem)) {
+            MessageUtils.throwError("afk.enchant_roller.error.is_not_axe");
+            ctx.stop();
+
+            return bl;
+        };
+
+        if (modConfig.preventAxeBreaking) {
+            int i = stack.getMaxDamage() - stack.getDamageValue();
+            if (i <= 10) {
+                MessageUtils.throwError("afk.enchant_roller.warn.low_damage");
+                ctx.stop();
+                bl = true;
             }
         }
 
-        return state;
+        return bl;
+    }
+
+    public static boolean swapItem(LocalPlayer player, EquipmentSlot handSlot, Predicate<Item> predicate) {
+        boolean success = false;
+
+        if (!predicate.test(player.getItemBySlot(handSlot).getItem())) {
+            Minecraft instance = Minecraft.getInstance();
+            if (instance == null) {
+                LOGGER.error("Minecraft instance is null, abort swap");
+                return false;
+            }
+
+            MultiPlayerGameMode interactionManager = instance.gameMode;
+            if (interactionManager == null) {
+                //? if > 1.21.1 {
+                /*LOGGER.error("GameMode null, player={}, slot={}", player.getGameProfile().name(), handSlot);
+                *///?} else {
+                LOGGER.error("GameMode null, player={}, slot={}", player.getGameProfile().getName(), handSlot);
+                //?}
+
+                return false;
+            };
+
+            ItemStack currentHandStack = player.getItemBySlot(handSlot);
+            if (predicate.test(currentHandStack.getItem())) {
+                return true;
+            }
+
+
+            if (player.inventoryMenu == player.containerMenu) {
+                success = swapFromInventory(player, interactionManager, predicate, handSlot);
+            } else {
+                success = swapFromHotbar(player, handSlot, predicate);
+            }
+        } else {
+            success = true;
+        }
+
+        return success;
     }
 
     private static boolean swapFromInventory(LocalPlayer player, MultiPlayerGameMode interactionManager, Predicate<Item> predicate, EquipmentSlot handSlot) {
-        boolean state = false;
+        boolean success = false;
 
-        for (int i = 9; i <= 44; i++) {
-            int slot = i >= 36 ? i - 36 : i;
-            var stack = player.getInventory().getItem(slot);
-            if (!stack.isEmpty() && predicate.test(stack.getItem())) {
-                boolean hasHandItem = !player.getItemBySlot(handSlot).isEmpty();
-                //here is mess.
-                int selectSlot = handSlot == EquipmentSlot.OFFHAND ? 45 : getSelect(player) + 36;
+        Inventory inventory = player.getInventory();
+        int targetSlot = handSlot == EquipmentSlot.OFFHAND ? PLAYER_INV_OFFHAND_CONTAINER_SLOT : getSelect(player) + PLAYER_INV_HOTBAR_CONTAINER_START;
 
+        for (int i = PLAYER_INV_STORAGE_START_SLOT; i < PLAYER_INV_STORAGE_END_SLOT; ++i) {
+            int slot = i >= PLAYER_INV_HOTBAR_CONTAINER_START ? i - PLAYER_INV_HOTBAR_CONTAINER_START : i;
+            var stack = inventory.getItem(slot);
+
+            if (stack.isEmpty() || !predicate.test(stack.getItem())) continue;
+
+            //here is mess.
+            boolean hasHandItem = !player.getItemBySlot(handSlot).isEmpty();
+            interactionManager.handleInventoryMouseClick(
+                    player.containerMenu.containerId,
+                    i,
+                    0,
+                    ClickType.PICKUP,
+                    player
+            );
+            interactionManager.handleInventoryMouseClick(
+                    player.containerMenu.containerId,
+                    targetSlot,
+                    0,
+                    ClickType.PICKUP,
+                    player
+            );
+
+            if (hasHandItem && inventory.getItem(slot).isEmpty()) {
                 interactionManager.handleInventoryMouseClick(
                         player.containerMenu.containerId,
                         i,
@@ -50,76 +141,60 @@ public class PlayerInventoryUtils {
                         ClickType.PICKUP,
                         player
                 );
-                interactionManager.handleInventoryMouseClick(
-                        player.containerMenu.containerId,
-                        selectSlot,
-                        0,
-                        ClickType.PICKUP,
-                        player
-                );
-
-                if (hasHandItem) {
-                    interactionManager.handleInventoryMouseClick(
-                            player.containerMenu.containerId,
-                            i,
-                            0,
-                            ClickType.PICKUP,
-                            player
-                    );
-                }
-
-                state = true;
-
-                break;
             }
+
+            //todo check here is saft?
+            player.containerMenu.broadcastChanges();
+
+            success = true;
         }
 
-        return state;
+        return success;
     }
 
     private static boolean swapFromHotbar(LocalPlayer player, EquipmentSlot handSlot, Predicate<Item> predicate) {
-        boolean state = false;
+        boolean success = false;
+        Inventory inventory = player.getInventory();
 
-        for (int i = 0; i <= 8; i++) {
-            var stack = player.getInventory().getItem(i);
-            if (!stack.isEmpty() && predicate.test(stack.getItem())) {
-                setSelect(player, i);
+        for (int i = 0; i < 9; ++i) {
+            var stack = inventory.getItem(i);
 
-                if (handSlot == EquipmentSlot.OFFHAND) {
-                    player.connection.send(
-                            new ServerboundPlayerActionPacket(
-                                    ServerboundPlayerActionPacket.Action.SWAP_ITEM_WITH_OFFHAND,
-                                    BlockPos.ZERO,
-                                    Direction.DOWN
-                            )
-                    );
-                }
+            if (stack.isEmpty() || !predicate.test(stack.getItem())) continue;
 
-                state = true;
+            setSelect(player, i);
 
-                break;
+            if (handSlot == EquipmentSlot.OFFHAND) {
+                player.connection.send(
+                        new ServerboundPlayerActionPacket(
+                                ServerboundPlayerActionPacket.Action.SWAP_ITEM_WITH_OFFHAND,
+                                BlockPos.ZERO,
+                                Direction.DOWN
+                        )
+                );
             }
+
+            success = true;
         }
 
-        return state;
+        return success;
     }
 
     public static int getSelect(LocalPlayer player) {
         //? if >=1.21.11 {
-        return player.getInventory().getSelectedSlot();
-        //?} else {
-        /*return player.getInventory().selected;
-        *///?}
+        /*return player.getInventory().getSelectedSlot();
+        *///?} else {
+        return player.getInventory().selected;
+        //?}
     }
 
     public static void setSelect(LocalPlayer player, int slot) {
         if (getSelect(player) == slot) return;
         //? if >=1.21.11 {
-        player.getInventory().setSelectedSlot(slot);
+        /*player.getInventory().setSelectedSlot(slot);
         player.connection.send(new ServerboundSetCarriedItemPacket(slot));
-        //?} else {
-        /*player.getInventory().selected = slot;
+        *///?} else {
+        player.getInventory().selected = slot;
         player.connection.send(new ServerboundSetCarriedItemPacket(slot));
-        *///?}
+        //?}
     }
 }