+++ CHANGELOG.md
@@ -1,1 +1,8 @@
-Fix: Fixed the issue where entering the world would crash.
+# 1.2.0
+this mod still under dev, so will get lots of bug
+break change: now you have to add enchantment with full translation key like minecraft:mending 1
+improve: remove some incomplete feature
+feat: add 3 new keys to do something
+- open config screen
+- add enchantment which in hand item
+- add enchantment which in inventory item(all item)